requirements.txt
python-telegram-bot==20.7
psycopg2-binary==2.9.9
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0
schedule==1.2.0
